add value, and some other tree properties
