+++ pyproject.toml
@@ -41,6 +41,8 @@ pytest-cov = "^2.10.1"
 sphinx-rtd-theme = "^0.5.1"
 pandas = "^1.0.3"
 pytest-xdist = "^2.2.1"
+#shap = "^0.39.0"
+#matplotlib = {version = "^3.4.2", python = ">=3.7,<4.0.0"}
 
 [build-system]
 requires = ["poetry-core>=1.0.0", "cython", "numpy"]
