+++ skgrf/base.py
@@ -36,7 +36,7 @@ class GRFMixin:
         """Validate cluster definitions against training data."""
         if cluster is None:
             self.classes_ = None
-            self.n_classes_ = None
+            self.n_classes_ = 1
             return np.array([])
         if len(cluster) != X.shape[0]:
             raise ValueError("cluster length must be the same as X")
@@ -142,10 +142,23 @@ class GRFMixin:
             )
 
     def _set_sample_weights(self, sample_weight):
+        # for accessing in Tree
         self.grf_forest_["leaf_weights"] = []
         for tree in self.grf_forest_["leaf_samples"]:
             self.grf_forest_["leaf_weights"].append([])
             for node in tree:
                 self.grf_forest_["leaf_weights"][-1].append(
-                    sum([sample_weight[idx] for idx in node])
+                    [sample_weight[idx] for idx in node]
                 )
+
+    def _set_leaf_values(self, y):
+        # for accessing in Tree
+        self.grf_forest_["leaf_values"] = []
+        for tree in self.grf_forest_["leaf_samples"]:
+            self.grf_forest_["leaf_values"].append([])
+            for node in tree:
+                self.grf_forest_["leaf_values"][-1].append([y[idx] for idx in node])
+
+    def _set_n_classes(self):
+        # for accessing in Tree
+        self.grf_forest_["n_classes"] = self.n_classes_
