+++ skgrf/ensemble/instrumental_regressor.py
@@ -215,6 +215,8 @@ class GRFInstrumentalRegressor(BaseGRFForest, RegressorMixin):
         self._set_sample_weights(
             sample_weight if sample_weight is not None else np.ones(len(X))
         )
+        self._set_leaf_values(y)
+        self._set_n_classes()
         return self
 
     def predict(self, X):
