+++ skgrf/ensemble/local_linear_regressor.py
@@ -203,6 +203,8 @@ class GRFLocalLinearRegressor(BaseGRFForest, RegressorMixin):
         self._set_sample_weights(
             sample_weight if sample_weight is not None else np.ones(len(X))
         )
+        self._set_leaf_values(y)
+        self._set_n_classes()
         return self
 
     def predict(self, X):
