+++ skgrf/ensemble/survival.py
@@ -164,6 +164,8 @@ class GRFSurvival(BaseGRFForest, BaseEstimator):
         self._set_sample_weights(
             sample_weight if sample_weight is not None else np.ones(len(X))
         )
+        self._set_leaf_values(y)
+        self._set_n_classes()
         return self
 
     def predict_cumulative_hazard_function(self, X):
