+++ skgrf/tree/instrumental_regressor.py
@@ -239,6 +239,8 @@ class GRFTreeInstrumentalRegressor(BaseGRFTree, RegressorMixin):
         self._set_sample_weights(
             sample_weight if sample_weight is not None else np.ones(len(X))
         )
+        self._set_leaf_values(y)
+        self._set_n_classes()
         return self
 
     def predict(self, X):
